refactor: 메시지 저장 트랜잭션에서 색인 이벤트 제거 및 아웃박스 패턴 적용
